requirements.txt
anthropic>=0.40.0
openai>=1.40.0
requests>=2.31.0
PyYAML>=6.0
